+++ blockchain/wallet.go
@@ -1,13 +1,13 @@
 package blockchain
 
-func (b *blockchain) BalanceOf(address string) int {
+func (b *blockchain) BalanceOf(address string) uint {
 	outputs := b.TxOutputsFrom(address)
 
 	return SumOverBalance(outputs)
 }
 
-func SumOverBalance(outputs []*TxOutput) int {
-	balance := 0
+func SumOverBalance(outputs []*TxOutput) uint {
+	balance := uint(0)
 
 	for _, output := range outputs {
 		balance += output.Amount
